+++ content/posts/index.mdx
@@ -8,7 +8,10 @@ img: ''
 tags:
   - welcome
   - wip
-title: Work. Progress? Next!
+  - ἐννεακαιδεκαετηρίς
+  - calendar
+subtitle: Progress? In. Work!
+title: Work. Progress! In?
 ---
 
 ## Welcome
@@ -17,9 +20,10 @@ In the late twentieth century, one man began a journey to hike his desk. This th
 
 ## Explore
 
-The site is still under heavy <i style="color: orange;" class="material-icons">construction</i>. Please bear with us as management wrestles the reticulating splines into subservient domination. Archived blog entries are to the left, grouped by Roman calendar year. [RSS](/rss.xml) is available as well.
+The site is still under heavy <i style="color: orange;" class="material-icons">construction</i>. Please bear with us as management wrestles the reticulating splines into subservient domination. Archived blog entries are to the left, grouped by [ἐννεακαιδεκαετηρίς](https://en.wikipedia.org/wiki/Metonic_cycle) year. For those feedsters, [RSS](/rss.xml) is available as well.
 
-Comments and other bloggy-mc-blogger style features will be coming in the next of the earth's rotations.
+- [ ] Comments and other bloggy-mc-blogger style features will be coming in the next of the earth's rotations.
+- [ ] Visualizations
 
 ## Recommendations for First-Time Readers
