--- src/components/Empty.tsx
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const Empty: React.FC = (): JSX.Element => <div />;
