+++ src/components/index.ts
@@ -1,17 +1,13 @@
 export * from './search';
 export * from './styles';
 export * from './AnchorTag';
-export * from './Cards';
-export * from './Chips';
 export * from './Comments';
-export * from './Empty';
 export * from './Header';
 export * from './Icon';
 export * from './Layout';
 export * from './LeftSidebar';
 export * from './Link';
 export * from './NextPrevious';
-export * from './NotFound';
-export * from './PageWrapper';
 export * from './RightSidebar';
+export * from './SEO';
 export * from './ThemeProvider';
